mysite/requirements.txt
Django
Djangorestframework
environs